Move IFileSystem/FileSystem -> Microsoft.VisualStudio.IO namespace
Utilities has become a grab bag of everything - it's not a great namespace name, and something we avoided using on the BCL.
