+++ src/Microsoft.VisualStudio.ProjectSystem.CSharp.VS/ProjectSystem/VS/Xproj/MigrateXprojProjectFactory.cs
@@ -1,15 +1,15 @@
 // Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
-using Microsoft.VisualStudio.Shell.Interop;
 using System;
-using Microsoft.VisualStudio.Packaging;
-using Microsoft.VisualStudio.Shell.Flavor;
-using System.Runtime.InteropServices;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using Microsoft.VisualStudio.ProjectSystem.Utilities;
+using System.Runtime.InteropServices;
 using System.Text;
+using Microsoft.VisualStudio.IO;
+using Microsoft.VisualStudio.Packaging;
+using Microsoft.VisualStudio.Shell.Flavor;
+using Microsoft.VisualStudio.Shell.Interop;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS.Xproj
 {
